Models: Fix normals on pedestal and unbroken wall
This commit fixes shading issues in the demo scene due to incorrect
"smooth" normals in the pedestal and unbroken wall.

The original models had non-orthogonal normals which caused the lighting
calculations to assume that the pedestal was meant to be an organic
shape along the edge between tops and sides. I did leave correctly
smooth normals around the cylindrical faces where appropriate.